fix scripts 03.sh

+++ ops/script/03.sh
@@ -13,14 +13,14 @@ caminho="./script_sample_folder/"
 if [ -d "$caminho" ]; then
     encontrado=false
     for pastas in "$caminho"*"$1"*; do
-        if [ -d "$pastas" ]; then
+        if [ $(ls -A $pastas) ] && [ -d "$pastas" ] ; then
             rm -r "$pastas"/*
             echo "Todos os arquivos em $pastas foram deletados."
             encontrado=true
         fi
     done
     if [ "$encontrado" = false ]; then
-        echo "Nenhuma pasta encontrada que corresponda ao padrão '$1'."
+        echo "Nenhuma pasta encontrada '$1'."
     fi
 else
     echo "A pasta $caminho não existe."
